Refactored 'Small Chirp' and 'Large Chirp' to 'Chirp Card'. Included comment display toggle button (can be modified further.)
