+++ resources/views/chirps/latest.blade.php
@@ -1,7 +1,7 @@
 <x-app-layout>
-    <div class="p-6" style="display:flex; flex-direction:column;width:550px;margin-top:2rem;">
+    <div class="p-6" style="display:flex; flex-direction:column;margin-top:2rem;">
        @foreach ($chirps as $chirp)
-            <x-chirp.small-chirp :chirp="$chirp"/>
+            <x-chirps.chirp-card :chirp="$chirp"/>
        @endforeach
     </div>
  </x-app-layout>
